# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ERPGo API
  description: |
    A comprehensive Enterprise Resource Planning (ERP) system built with Go, providing user management, inventory management, order management, and business analytics capabilities.

    ## Overview
    ERPGo is a modern, cloud-native ERP system designed for high performance, scalability, and maintainability. It provides:

    - User authentication and authorization with JWT tokens
    - Product catalog management with categories and variants
    - Inventory tracking and warehouse management
    - Order processing and fulfillment workflows
    - Customer management and address handling
    - Comprehensive analytics and reporting

    ## Base URLs
    - **Production**: `https://api.erpgo.example.com`
    - **Staging**: `https://api-staging.erpgo.example.com`
    - **Development**: `http://localhost:8080`

    ## Authentication
    The API uses JWT (JSON Web Tokens) for authentication. Include the token in the `Authorization` header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    API requests are rate-limited to ensure fair usage:
    - **Default**: 100 requests per minute
    - **Burst**: 200 requests
    - **Window**: 1 minute

    Rate limit headers are included in responses:
    ```
    X-RateLimit-Limit: 100
    X-RateLimit-Remaining: 99
    X-RateLimit-Reset: 1640995200
    ```
  version: 1.0.0
  contact:
    name: ERPGo Support
    email: support@erpgo.example.com
...
  license:
...
servers:
...
paths:
...
  /api/v1/categories:
    post:
      tags:
        - Categories
      summary: Create category
      description: Create a new product category
      operationId: createCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'

    get:
      tags:
        - Categories
      summary: List categories
      description: Get list of categories
      operationId: listCategories
      security:
        - bearerAuth: []
      parameters:
        - name: parent_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by parent category ID
        - name: is_active
          in: query
          schema:
            type: boolean
          description: Filter by active status
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
  /api/v1/categories/tree:
    get:
      tags:
        - Categories
      summary: Get category tree
      description: Get hierarchical tree of categories
      operationId: getCategoryTree
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Category tree retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryTreeResponse'

  # Order management endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "Electronics"
        description:
          type: string
          example: "Electronic devices and accessories"
        parent_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        parent:
          $ref: '#/components/schemas/CategoryInfo'
        path:
          type: string
          example: "Electronics/Smartphones"
        level:
          type: integer
          example: 2
        is_active:
          type: boolean
          example: true
        sort_order:
          type: integer
          example: 1
        product_count:
          type: integer
          example: 25
        created_at:
          type: string
          format: date-time
          example: "2024-01-08T10:30:00Z"
    CreateCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Electronics"
        description:
          type: string
          example: "Electronic devices and accessories"
        parent_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        sort_order:
          type: integer
          default: 0
    CategoryInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        path:
          type: string
    CategoryListResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CategoryTreeResponse:
      type: object
      properties:
        tree:
          type: array
          items:
            $ref: '#/components/schemas/CategoryTreeNode'
    CategoryTreeNode:
      type: object
      allOf:
        - $ref: '#/components/schemas/CategoryResponse'
        - type: object
          properties:
            children:
              type: array
              items:
                $ref: '#/components/schemas/CategoryTreeNode'

    # Order schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
...